Changes for order of the exercises and renaming
